Add files via upload
Created a basic algorithm to find reorder quantity (Only did first chunk: 1 - 15 week window). (Still trying to figure out if I want to do a 2 - 16 week window next or a 15 - 30). Have to come back to this and implement my own changes in the future for the algorithm to see if it can improve.